# 8310268: RISC-V: misaligned memory access in String.Compare intrinsic