--- a/src/hotspot/cpu/riscv/stubGenerator_riscv.cpp
+++ b/src/hotspot/cpu/riscv/stubGenerator_riscv.cpp
@@ -2310,24 +2310,23 @@ class StubGenerator: public StubCodeGenerator {
   }
 
   // code for comparing 8 characters of strings with Latin1 and Utf16 encoding
-  void compare_string_8_x_LU(Register tmpL, Register tmpU, Label &DIFF1,
-                              Label &DIFF2) {
-    const Register strU = x12, curU = x7, strL = x29, tmp = x30;
+  void compare_string_8_x_LU(Register tmpL, Register tmpU, Register strL, Register strU, Label& DIFF) {
+    const Register tmp = x30;
     __ ld(tmpL, Address(strL));
-    __ addi(strL, strL, 8);
+    __ addi(strL, strL, wordSize);
     __ ld(tmpU, Address(strU));
-    __ addi(strU, strU, 8);
+    __ addi(strU, strU, wordSize);
     __ inflate_lo32(tmp, tmpL);
     __ mv(t0, tmp);
-    __ xorr(tmp, curU, t0);
-    __ bnez(tmp, DIFF2);
+    __ xorr(tmp, tmpU, t0);
+    __ bnez(tmp, DIFF);
 
-    __ ld(curU, Address(strU));
-    __ addi(strU, strU, 8);
+    __ ld(tmpU, Address(strU));
+    __ addi(strU, strU, wordSize);
     __ inflate_hi32(tmp, tmpL);
     __ mv(t0, tmp);
     __ xorr(tmp, tmpU, t0);
-    __ bnez(tmp, DIFF1);
+    __ bnez(tmp, DIFF);
   }
 
   // x10  = result
@@ -2342,8 +2341,7 @@ class StubGenerator: public StubCodeGenerator {
     __ align(CodeEntryAlignment);
     StubCodeMark mark(this, "StubRoutines", isLU ? "compare_long_string_different_encoding LU" : "compare_long_string_different_encoding UL");
     address entry = __ pc();
-    Label SMALL_LOOP, TAIL, TAIL_LOAD_16, LOAD_LAST, DIFF1, DIFF2,
-          DONE, CALCULATE_DIFFERENCE;
+    Label SMALL_LOOP, TAIL, LOAD_LAST, DIFF, DONE, CALCULATE_DIFFERENCE;
     const Register result = x10, str1 = x11, cnt1 = x12, str2 = x13, cnt2 = x14,
                    tmp1 = x28, tmp2 = x29, tmp3 = x30, tmp4 = x7, tmp5 = x31;
     RegSet spilled_regs = RegSet::of(tmp4, tmp5);
@@ -2354,16 +2352,9 @@ class StubGenerator: public StubCodeGenerator {
     __ mv(isLU ? tmp1 : tmp2, tmp3);
     __ addi(str1, str1, isLU ? wordSize / 2 : wordSize);
     __ addi(str2, str2, isLU ? wordSize : wordSize / 2);
-    __ sub(cnt2, cnt2, 8); // Already loaded 4 symbols. Last 4 is special case.
+    __ sub(cnt2, cnt2, wordSize / 2); // Already loaded 4 symbols
     __ push_reg(spilled_regs, sp);
 
-    if (isLU) {
-      __ add(str1, str1, cnt2);
-      __ shadd(str2, cnt2, str2, t0, 1);
-    } else {
-      __ shadd(str1, cnt2, str1, t0, 1);
-      __ add(str2, str2, cnt2);
-    }
     __ xorr(tmp3, tmp1, tmp2);
     __ mv(tmp5, tmp2);
     __ bnez(tmp3, CALCULATE_DIFFERENCE);
@@ -2373,46 +2364,93 @@ class StubGenerator: public StubCodeGenerator {
              tmpU = isLU ? tmp5 : tmp1, // where to keep U for comparison
              tmpL = isLU ? tmp1 : tmp5; // where to keep L for comparison
 
-    __ sub(tmp2, strL, cnt2); // strL pointer to load from
-    __ slli(t0, cnt2, 1);
-    __ sub(cnt1, strU, t0); // strU pointer to load from
+    // make sure main loop is 8 byte-aligned, we should load another 4 bytes from strL
+    __ beqz(cnt2, DONE);  // no characters left
+    __ lwu(tmpL, Address(strL));
+    __ addi(strL, strL, wordSize / 2);
+    __ ld(tmpU, Address(strU));
+    __ addi(strU, strU, wordSize);
+    __ inflate_lo32(tmp3, tmpL);
+    __ mv(tmpL, tmp3);
+    __ xorr(tmp3, tmpU, tmpL);
+    __ bnez(tmp3, CALCULATE_DIFFERENCE);
+    __ addi(cnt2, cnt2, -wordSize / 2);
 
-    __ ld(tmp4, Address(cnt1));
-    __ addi(cnt1, cnt1, 8);
-    __ beqz(cnt2, LOAD_LAST); // no characters left except last load
-    __ sub(cnt2, cnt2, 16);
+    __ beqz(cnt2, DONE);  // no character left
+    // we are now 8-bytes aligned on strL
+    __ sub(cnt2, cnt2, wordSize * 2);
     __ bltz(cnt2, TAIL);
     __ bind(SMALL_LOOP); // smaller loop
-      __ sub(cnt2, cnt2, 16);
-      compare_string_8_x_LU(tmpL, tmpU, DIFF1, DIFF2);
-      compare_string_8_x_LU(tmpL, tmpU, DIFF1, DIFF2);
+      __ sub(cnt2, cnt2, wordSize * 2);
+      compare_string_8_x_LU(tmpL, tmpU, strL, strU, DIFF);
+      compare_string_8_x_LU(tmpL, tmpU, strL, strU, DIFF);
       __ bgez(cnt2, SMALL_LOOP);
-      __ addi(t0, cnt2, 16);
-      __ beqz(t0, LOAD_LAST);
-    __ bind(TAIL); // 1..15 characters left until last load (last 4 characters)
-      // Address of 8 bytes before last 4 characters in UTF-16 string
-      __ shadd(cnt1, cnt2, cnt1, t0, 1);
-      // Address of 16 bytes before last 4 characters in Latin1 string
-      __ add(tmp2, tmp2, cnt2);
-      __ ld(tmp4, Address(cnt1, -8));
-      // last 16 characters before last load
-      compare_string_8_x_LU(tmpL, tmpU, DIFF1, DIFF2);
-      compare_string_8_x_LU(tmpL, tmpU, DIFF1, DIFF2);
-      __ j(LOAD_LAST);
-    __ bind(DIFF2);
-      __ mv(tmpU, tmp4);
-    __ bind(DIFF1);
+      __ addi(t0, cnt2, wordSize * 2);
+      __ beqz(t0, DONE);
+    __ bind(TAIL);  // 1..15 characters left
+    if (AvoidUnalignedAccesses) {
+      // Aligned access. Load bytes in portions - 4, 2, 1.
+      Label LOAD_LAST, WORD_CMP, TAIL01, TAIL02;
+
+      __ addi(t0, cnt2, wordSize);
+      __ addi(cnt2, cnt2, wordSize*2); // amount of characters left to process
+      __ bltz(t0, LOAD_LAST);
+      // remaining characters are greater than or equals to 8, we can do one compare_string_8_x_LU
+      compare_string_8_x_LU(tmpL, tmpU, strL, strU, DIFF);
+      __ addi(cnt2, cnt2, -wordSize);
+      __ beqz(cnt2, DONE);  // no character left
+      __ bind(LOAD_LAST);   // cnt2 = 1..7 characters left
+      // load four bytes if (cnt2 & 4) is not 0
+      __ test_bit(t0, cnt2, 2);
+      __ beqz(t0, TAIL02);
+      {
+        __ lwu(tmpL, Address(strL));
+        __ addi(strL, strL, wordSize / 2);
+        __ ld(tmpU, Address(strU));
+        __ addi(strU, strU, wordSize);
+        __ inflate_lo32(tmp3, tmpL);
+        __ mv(tmpL, tmp3);
+        __ xorr(tmp3, tmpU, tmpL);
+        __ bnez(tmp3, CALCULATE_DIFFERENCE);
+        __ addi(cnt2, cnt2, -wordSize / 2);
+      }
+      __ bind(TAIL02);
+      // load two bytes if (cnt2 & 2) is not 0
+      __ test_bit(t0, cnt2, 1);
+      __ beqz(t0, TAIL01);
+      {
+        __ lhu(tmpL, Address(strL));
+        __ addi(strL, strL, wordSize / 4);
+        __ lwu(tmpU, Address(strU));
+        __ addi(strU, strU, wordSize / 2);
+        __ inflate_lo32(tmp3, tmpL);
+        __ mv(tmpL, tmp3);
+        __ xorr(tmp3, tmpU, tmpL);
+        __ bnez(tmp3, CALCULATE_DIFFERENCE);
+        __ addi(cnt2, cnt2, -wordSize / 4);
+      }
+      __ bind(TAIL01);
+      // load one byte if anything left in cnt2
+      __ beqz(cnt2, DONE);
+      {
+        __ lbu(tmpL, Address(strL)); // when only one byte left - no need to inflate
+        __ lhu(tmpU, Address(strU));
+        __ xorr(tmp3, tmpU, tmpL);
+        __ bnez(tmp3, CALCULATE_DIFFERENCE);
+      }
+      __ j(DONE); // no character left
+    } else {
+      // Unaligned accesses. Load from non-byte aligned address.
+      __ slli(t0, cnt2, 1);     // now in bytes
+      __ add(strU, strU, t0);   // Address of last 8 bytes in UTF-16 string
+      __ add(strL, strL, cnt2); // Address of last 16 bytes in Latin1 string
+      // last 16 characters
+      compare_string_8_x_LU(tmpL, tmpU, strL, strU, DIFF);
+      compare_string_8_x_LU(tmpL, tmpU, strL, strU, DIFF);
+      __ j(DONE);
+    }
+    __ bind(DIFF);
       __ mv(tmpL, t0);
-      __ j(CALCULATE_DIFFERENCE);
-    __ bind(LOAD_LAST);
-      // Last 4 UTF-16 characters are already pre-loaded into tmp4 by compare_string_8_x_LU.
-      // No need to load it again
-      __ mv(tmpU, tmp4);
-      __ ld(tmpL, Address(strL));
-      __ inflate_lo32(tmp3, tmpL);
-      __ mv(tmpL, tmp3);
-      __ xorr(tmp3, tmpU, tmpL);
-      __ beqz(tmp3, DONE);
 
       // Find the first different characters in the longwords and
       // compute their difference.
@@ -2529,6 +2567,7 @@ class StubGenerator: public StubCodeGenerator {
       __ addi(str1, str1, 8);
       __ ld(tmp2, Address(str2));
       __ addi(str2, str2, 8);
+      __ beqz(cnt2, LAST_CHECK_AND_LENGTH_DIFF);
       __ sub(cnt2, cnt2, isLL ? 8 : 4);
     __ bind(CHECK_LAST);
       if (!isLL) {
@@ -2537,9 +2576,9 @@ class StubGenerator: public StubCodeGenerator {
       __ xorr(tmp4, tmp1, tmp2);
       __ bnez(tmp4, DIFF);
       __ add(str1, str1, cnt2);
-      __ ld(tmp5, Address(str1));
+      __ load_long_misaligned(tmp5, Address(str1), tmp3, isLL ? 1 : 2);
       __ add(str2, str2, cnt2);
-      __ ld(cnt1, Address(str2));
+      __ load_long_misaligned(cnt1, Address(str2), tmp3, isLL ? 1 : 2);
       __ xorr(tmp4, tmp5, cnt1);
       __ beqz(tmp4, LENGTH_DIFF);
       // Find the first different characters in the longwords and