… test/hotspot/jtreg/compiler/intrinsics/string/TestStringCompareToDifferentLength.java