Please review this fix. it eliminates misaligned loads in String.compare on risc-v

for small compares ( <= 72 bytes), the instrinsic in c2_MacroAssembler.cpp is used,
it reads ( in case of UU/LL) 8 bytes per loop, and at then end, it reads tail - misaligned load of last 8 bytes from the string.

so if string length is not 8x bytes long then last load is misaligned, also it performs read/compare of some data which already was processed.

I have changed that to compare only last length%8 bytes using SHORT_STRING part of intrinsic for UL/LU. But for UU/LL I have made an optimised version.

Thanks to optimisations for conditional branching at line [947](https://github.com/openjdk/jdk/pull/14534/files#diff-35eb1d2f1e2f0514dd46bd7fbad49ff2c87703d5a3041a6433956df00a3fe6e6R947) I’ve got no perf drop on thead ( with +AvoidUnalignedAccesses) which supports misaligned access.

Improvements to inflate_XX() methods gives 3-5% improvements for UL/LU cases on thead, almost no perf change on hifive.

for large strings, the instrinsics from stubGenerator.cpp is used
for UU/LL - generate_compare_long_string_same_encoding, I have just replaced misaligned ld with load_long_misaligned. Since this tail reading is not on hot path, this give some small penalty for thead when -XX:+AvoidUnalignedAccesses.

large LU/UL comparision is done in compare_long_string_different_encoding in sutbGenerator.cpp:
These changes are partially based on feilongjiang's patch, but I have changed tail reading to prevent reading past the end of string. I have observed no perf difference between feilongjiang's and my version.

This also enables regression test for string.Compare which previously was aarch64-only

Testing: tier1 and tier2 clean on hifive.

JMH testing, hifive:
before:
```
Benchmark                                   (delta)  (size)  Mode  Cnt     Score    Error  Units
StringCompareToDifferentLength.compareToLL        2      24  avgt    9     6.474 ±  1.475  ms/op
StringCompareToDifferentLength.compareToLL        2      36  avgt    9   125.823 ±  1.947  ms/op
StringCompareToDifferentLength.compareToLL        2      72  avgt    9    10.512 ±  0.236  ms/op
StringCompareToDifferentLength.compareToLL        2     128  avgt    9    13.032 ±  0.821  ms/op
StringCompareToDifferentLength.compareToLL        2     256  avgt    9    18.983 ±  0.318  ms/op
StringCompareToDifferentLength.compareToLL        2     512  avgt    9    29.925 ±  0.458  ms/op
StringCompareToDifferentLength.compareToLL        2     520  avgt    9    30.607 ±  0.963  ms/op
StringCompareToDifferentLength.compareToLL        2     523  avgt    9   150.772 ±  1.048  ms/op
StringCompareToDifferentLength.compareToLU        2      24  avgt    9    13.979 ±  0.868  ms/op
StringCompareToDifferentLength.compareToLU        2      36  avgt    9    18.344 ±  0.640  ms/op
StringCompareToDifferentLength.compareToLU        2      72  avgt    9   567.667 ±  5.222  ms/op
StringCompareToDifferentLength.compareToLU        2     128  avgt    9  1060.860 ±  3.348  ms/op
StringCompareToDifferentLength.compareToLU        2     256  avgt    9  2054.307 ±  6.988  ms/op
StringCompareToDifferentLength.compareToLU        2     512  avgt    9  4025.110 ± 16.923  ms/op
StringCompareToDifferentLength.compareToLU        2     520  avgt    9  4030.565 ± 18.757  ms/op
StringCompareToDifferentLength.compareToLU        2     523  avgt    9  4447.476 ± 21.069  ms/op
StringCompareToDifferentLength.compareToUL        2      24  avgt    9    16.458 ±  0.640  ms/op
StringCompareToDifferentLength.compareToUL        2      36  avgt    9    20.713 ±  0.705  ms/op
StringCompareToDifferentLength.compareToUL        2      72  avgt    9   566.969 ±  3.499  ms/op
StringCompareToDifferentLength.compareToUL        2     128  avgt    9  1061.740 ±  3.034  ms/op
StringCompareToDifferentLength.compareToUL        2     256  avgt    9  2050.731 ±  6.786  ms/op
StringCompareToDifferentLength.compareToUL        2     512  avgt    9  4028.398 ± 16.138  ms/op
StringCompareToDifferentLength.compareToUL        2     520  avgt    9  4034.976 ± 19.330  ms/op
StringCompareToDifferentLength.compareToUL        2     523  avgt    9  4445.266 ± 13.599  ms/op
StringCompareToDifferentLength.compareToUU        2      24  avgt    9     7.673 ±  0.933  ms/op
StringCompareToDifferentLength.compareToUU        2      36  avgt    9     9.495 ±  1.561  ms/op
StringCompareToDifferentLength.compareToUU        2      72  avgt    9    14.430 ±  0.370  ms/op
StringCompareToDifferentLength.compareToUU        2     128  avgt    9    19.752 ±  0.927  ms/op
StringCompareToDifferentLength.compareToUU        2     256  avgt    9    30.463 ±  0.770  ms/op
StringCompareToDifferentLength.compareToUU        2     512  avgt    9    52.556 ±  0.588  ms/op
StringCompareToDifferentLength.compareToUU        2     520  avgt    9    53.675 ±  0.544  ms/op
StringCompareToDifferentLength.compareToUU        2     523  avgt    9   173.248 ±  0.830  ms/op
```
after:
```
Benchmark                                   (delta)  (size)  Mode  Cnt    Score    Error  Units
StringCompareToDifferentLength.compareToLL        2      24  avgt    9    6.403 ±  1.252  ms/op
StringCompareToDifferentLength.compareToLL        2      36  avgt    9    8.473 ±  1.084  ms/op
StringCompareToDifferentLength.compareToLL        2      72  avgt    9   11.301 ±  1.067  ms/op
StringCompareToDifferentLength.compareToLL        2     128  avgt    9   15.615 ±  0.825  ms/op
StringCompareToDifferentLength.compareToLL        2     256  avgt    9   21.007 ±  0.683  ms/op
StringCompareToDifferentLength.compareToLL        2     512  avgt    9   32.836 ±  0.639  ms/op
StringCompareToDifferentLength.compareToLL        2     520  avgt    9   30.565 ±  1.188  ms/op
StringCompareToDifferentLength.compareToLL        2     523  avgt    9   33.284 ±  0.862  ms/op
StringCompareToDifferentLength.compareToLU        2      24  avgt    9   14.089 ±  0.563  ms/op
StringCompareToDifferentLength.compareToLU        2      36  avgt    9   17.754 ±  0.622  ms/op
StringCompareToDifferentLength.compareToLU        2      72  avgt    9   32.248 ±  0.679  ms/op
StringCompareToDifferentLength.compareToLU        2     128  avgt    9   52.433 ±  4.008  ms/op
StringCompareToDifferentLength.compareToLU        2     256  avgt    9   91.470 ±  0.927  ms/op
StringCompareToDifferentLength.compareToLU        2     512  avgt    9  174.563 ±  1.376  ms/op
StringCompareToDifferentLength.compareToLU        2     520  avgt    9  178.459 ±  1.381  ms/op
StringCompareToDifferentLength.compareToLU        2     523  avgt    9  202.416 ± 13.131  ms/op
StringCompareToDifferentLength.compareToUL        2      24  avgt    9   15.420 ±  0.192  ms/op
StringCompareToDifferentLength.compareToUL        2      36  avgt    9   19.283 ±  0.226  ms/op
StringCompareToDifferentLength.compareToUL        2      72  avgt    9   34.827 ±  1.365  ms/op
StringCompareToDifferentLength.compareToUL        2     128  avgt    9   52.482 ±  1.852  ms/op
StringCompareToDifferentLength.compareToUL        2     256  avgt    9   94.261 ±  1.677  ms/op
StringCompareToDifferentLength.compareToUL        2     512  avgt    9  179.576 ±  3.538  ms/op
StringCompareToDifferentLength.compareToUL        2     520  avgt    9  179.983 ±  1.184  ms/op
StringCompareToDifferentLength.compareToUL        2     523  avgt    9  180.987 ±  2.206  ms/op
StringCompareToDifferentLength.compareToUU        2      24  avgt    9   10.841 ±  0.905  ms/op
StringCompareToDifferentLength.compareToUU        2      36  avgt    9   11.709 ±  0.517  ms/op
StringCompareToDifferentLength.compareToUU        2      72  avgt    9   15.979 ±  0.691  ms/op
StringCompareToDifferentLength.compareToUU        2     128  avgt    9   20.225 ±  0.623  ms/op
StringCompareToDifferentLength.compareToUU        2     256  avgt    9   31.435 ±  1.017  ms/op
StringCompareToDifferentLength.compareToUU        2     512  avgt    9   53.657 ±  1.057  ms/op
StringCompareToDifferentLength.compareToUU        2     520  avgt    9   53.921 ±  1.079  ms/op
StringCompareToDifferentLength.compareToUU        2     523  avgt    9   54.215 ±  0.655  ms/op
```

thead:

before:
```
Benchmark                                   (delta)  (size)  Mode  Cnt    Score   Error  Units
StringCompareToDifferentLength.compareToLL        2      24  avgt    9    2.592 ? 0.008  ms/op
StringCompareToDifferentLength.compareToLL        2      36  avgt    9    3.727 ? 0.115  ms/op
StringCompareToDifferentLength.compareToLL        2      72  avgt    9    7.277 ? 0.076  ms/op
StringCompareToDifferentLength.compareToLL        2     128  avgt    9    8.733 ? 0.178  ms/op
StringCompareToDifferentLength.compareToLL        2     256  avgt    9   13.334 ? 0.448  ms/op
StringCompareToDifferentLength.compareToLL        2     512  avgt    9   19.843 ? 0.234  ms/op
StringCompareToDifferentLength.compareToLL        2     520  avgt    9   19.971 ? 0.093  ms/op
StringCompareToDifferentLength.compareToLL        2     523  avgt    9   21.052 ? 0.318  ms/op
StringCompareToDifferentLength.compareToLU        2      24  avgt    9    7.847 ? 0.188  ms/op
StringCompareToDifferentLength.compareToLU        2      36  avgt    9   10.890 ? 0.077  ms/op
StringCompareToDifferentLength.compareToLU        2      72  avgt    9   24.353 ? 0.277  ms/op
StringCompareToDifferentLength.compareToLU        2     128  avgt    9   40.850 ? 0.591  ms/op
StringCompareToDifferentLength.compareToLU        2     256  avgt    9   72.205 ? 0.352  ms/op
StringCompareToDifferentLength.compareToLU        2     512  avgt    9  134.445 ? 0.523  ms/op
StringCompareToDifferentLength.compareToLU        2     520  avgt    9  134.866 ? 1.081  ms/op
StringCompareToDifferentLength.compareToLU        2     523  avgt    9  138.925 ? 1.241  ms/op
StringCompareToDifferentLength.compareToUL        2      24  avgt    9    9.532 ? 0.080  ms/op
StringCompareToDifferentLength.compareToUL        2      36  avgt    9   12.873 ? 0.304  ms/op
StringCompareToDifferentLength.compareToUL        2      72  avgt    9   26.725 ? 0.582  ms/op
StringCompareToDifferentLength.compareToUL        2     128  avgt    9   42.855 ? 0.451  ms/op
StringCompareToDifferentLength.compareToUL        2     256  avgt    9   74.674 ? 0.817  ms/op
StringCompareToDifferentLength.compareToUL        2     512  avgt    9  136.531 ? 0.530  ms/op
StringCompareToDifferentLength.compareToUL        2     520  avgt    9  136.446 ? 0.946  ms/op
StringCompareToDifferentLength.compareToUL        2     523  avgt    9  141.521 ? 1.371  ms/op
StringCompareToDifferentLength.compareToUU        2      24  avgt    9    4.106 ? 0.012  ms/op
StringCompareToDifferentLength.compareToUU        2      36  avgt    9    5.458 ? 0.237  ms/op
StringCompareToDifferentLength.compareToUU        2      72  avgt    9    9.530 ? 0.155  ms/op
StringCompareToDifferentLength.compareToUU        2     128  avgt    9   13.908 ? 0.390  ms/op
StringCompareToDifferentLength.compareToUU        2     256  avgt    9   20.331 ? 0.221  ms/op
StringCompareToDifferentLength.compareToUU        2     512  avgt    9   33.734 ? 0.213  ms/op
StringCompareToDifferentLength.compareToUU        2     520  avgt    9   33.884 ? 0.201  ms/op
StringCompareToDifferentLength.compareToUU        2     523  avgt    9   35.376 ? 0.210  ms/op
```

after:
```
Benchmark                                   (delta)  (size)  Mode  Cnt    Score   Error  Units
StringCompareToDifferentLength.compareToLL        2      24  avgt    9    2.701 ? 0.128  ms/op
StringCompareToDifferentLength.compareToLL        2      36  avgt    9    3.552 ? 0.024  ms/op
StringCompareToDifferentLength.compareToLL        2      72  avgt    9    7.533 ? 0.049  ms/op
StringCompareToDifferentLength.compareToLL        2     128  avgt    9   10.330 ? 0.090  ms/op
StringCompareToDifferentLength.compareToLL        2     256  avgt    9   15.484 ? 0.474  ms/op
StringCompareToDifferentLength.compareToLL        2     512  avgt    9   21.892 ? 0.366  ms/op
StringCompareToDifferentLength.compareToLL        2     520  avgt    9   20.239 ? 0.102  ms/op
StringCompareToDifferentLength.compareToLL        2     523  avgt    9   22.113 ? 0.302  ms/op
StringCompareToDifferentLength.compareToLU        2      24  avgt    9    7.150 ? 0.065  ms/op
StringCompareToDifferentLength.compareToLU        2      36  avgt    9   10.103 ? 0.227  ms/op
StringCompareToDifferentLength.compareToLU        2      72  avgt    9   22.354 ? 0.218  ms/op
StringCompareToDifferentLength.compareToLU        2     128  avgt    9   36.192 ? 0.374  ms/op
StringCompareToDifferentLength.compareToLU        2     256  avgt    9   65.506 ? 0.301  ms/op
StringCompareToDifferentLength.compareToLU        2     512  avgt    9  124.193 ? 0.337  ms/op
StringCompareToDifferentLength.compareToLU        2     520  avgt    9  126.693 ? 0.560  ms/op
StringCompareToDifferentLength.compareToLU        2     523  avgt    9  127.654 ? 0.598  ms/op
StringCompareToDifferentLength.compareToUL        2      24  avgt    9    9.611 ? 0.558  ms/op
StringCompareToDifferentLength.compareToUL        2      36  avgt    9   12.892 ? 0.672  ms/op
StringCompareToDifferentLength.compareToUL        2      72  avgt    9   23.973 ? 0.570  ms/op
StringCompareToDifferentLength.compareToUL        2     128  avgt    9   38.567 ? 1.064  ms/op
StringCompareToDifferentLength.compareToUL        2     256  avgt    9   67.992 ? 0.986  ms/op
StringCompareToDifferentLength.compareToUL        2     512  avgt    9  126.910 ? 0.738  ms/op
StringCompareToDifferentLength.compareToUL        2     520  avgt    9  128.019 ? 0.508  ms/op
StringCompareToDifferentLength.compareToUL        2     523  avgt    9  127.973 ? 0.326  ms/op
StringCompareToDifferentLength.compareToUU        2      24  avgt    9    3.903 ? 0.020  ms/op
StringCompareToDifferentLength.compareToUU        2      36  avgt    9    4.911 ? 0.044  ms/op
StringCompareToDifferentLength.compareToUU        2      72  avgt    9   10.019 ? 0.210  ms/op
StringCompareToDifferentLength.compareToUU        2     128  avgt    9   14.479 ? 0.217  ms/op
StringCompareToDifferentLength.compareToUU        2     256  avgt    9   20.668 ? 0.208  ms/op
StringCompareToDifferentLength.compareToUU        2     512  avgt    9   34.244 ? 0.240  ms/op
StringCompareToDifferentLength.compareToUU        2     520  avgt    9   34.582 ? 0.220  ms/op
StringCompareToDifferentLength.compareToUU        2     523  avgt    9   34.804 ? 0.199  ms/op
```

after, with -XX:-AvoidUnalignedAccesses:
```
Benchmark                                   (delta)  (size)  Mode  Cnt    Score   Error  Units
StringCompareToDifferentLength.compareToLL        2      24  avgt    9    2.641 ? 0.221  ms/op
StringCompareToDifferentLength.compareToLL        2      36  avgt    9    3.486 ? 0.047  ms/op
StringCompareToDifferentLength.compareToLL        2      72  avgt    9    7.528 ? 0.037  ms/op
StringCompareToDifferentLength.compareToLL        2     128  avgt    9    9.002 ? 0.037  ms/op
StringCompareToDifferentLength.compareToLL        2     256  avgt    9   13.772 ? 0.185  ms/op
StringCompareToDifferentLength.compareToLL        2     512  avgt    9   20.139 ? 0.138  ms/op
StringCompareToDifferentLength.compareToLL        2     520  avgt    9   20.238 ? 0.322  ms/op
StringCompareToDifferentLength.compareToLL        2     523  avgt    9   20.928 ? 0.171  ms/op
StringCompareToDifferentLength.compareToLU        2      24  avgt    9    7.100 ? 0.113  ms/op
StringCompareToDifferentLength.compareToLU        2      36  avgt    9    9.976 ? 0.213  ms/op
StringCompareToDifferentLength.compareToLU        2      72  avgt    9   21.896 ? 0.205  ms/op
StringCompareToDifferentLength.compareToLU        2     128  avgt    9   37.569 ? 0.113  ms/op
StringCompareToDifferentLength.compareToLU        2     256  avgt    9   66.993 ? 0.730  ms/op
StringCompareToDifferentLength.compareToLU        2     512  avgt    9  126.249 ? 2.044  ms/op
StringCompareToDifferentLength.compareToLU        2     520  avgt    9  126.366 ? 0.585  ms/op
StringCompareToDifferentLength.compareToLU        2     523  avgt    9  129.772 ? 0.538  ms/op
StringCompareToDifferentLength.compareToUL        2      24  avgt    9   10.031 ? 0.305  ms/op
StringCompareToDifferentLength.compareToUL        2      36  avgt    9   12.185 ? 0.629  ms/op
StringCompareToDifferentLength.compareToUL        2      72  avgt    9   24.069 ? 0.539  ms/op
StringCompareToDifferentLength.compareToUL        2     128  avgt    9   40.073 ? 0.573  ms/op
StringCompareToDifferentLength.compareToUL        2     256  avgt    9   69.349 ? 0.609  ms/op
StringCompareToDifferentLength.compareToUL        2     512  avgt    9  128.168 ? 0.843  ms/op
StringCompareToDifferentLength.compareToUL        2     520  avgt    9  128.314 ? 0.584  ms/op
StringCompareToDifferentLength.compareToUL        2     523  avgt    9  132.530 ? 0.373  ms/op
StringCompareToDifferentLength.compareToUU        2      24  avgt    9    3.658 ? 0.160  ms/op
StringCompareToDifferentLength.compareToUU        2      36  avgt    9    4.496 ? 0.026  ms/op
StringCompareToDifferentLength.compareToUU        2      72  avgt    9    9.397 ? 0.116  ms/op
StringCompareToDifferentLength.compareToUU        2     128  avgt    9   13.752 ? 0.409  ms/op
StringCompareToDifferentLength.compareToUU        2     256  avgt    9   20.053 ? 0.317  ms/op
StringCompareToDifferentLength.compareToUU        2     512  avgt    9   33.385 ? 0.242  ms/op
StringCompareToDifferentLength.compareToUU        2     520  avgt    9   33.808 ? 0.301  ms/op
StringCompareToDifferentLength.compareToUU        2     523  avgt    9   35.481 ? 0.480  ms/op

```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8310268](https://bugs.openjdk.org/browse/JDK-8310268): RISC-V: misaligned memory access in String.Compare intrinsic (**Enhancement** - P4)


### Reviewers
 * [Fei Yang](https://openjdk.org/census#fyang) (@RealFYang - **Reviewer**) ⚠️ Review applies to [f6e066e5](https://git.openjdk.org/jdk/pull/14534/files/f6e066e58650f9a324fed3d76c236e2c8a75c004)


### Contributors
 * Feilong Jiang `<fjiang@openjdk.org>`

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/14534/head:pull/14534` \
`$ git checkout pull/14534`

Update a local copy of the PR: \
`$ git checkout pull/14534` \
`$ git pull https://git.openjdk.org/jdk.git pull/14534/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 14534`

View PR using the GUI difftool: \
`$ git pr show -t 14534`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/14534.diff">https://git.openjdk.org/jdk/pull/14534.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/14534#issuecomment-1596559700)