@RealFYang 2023-06-26
Hi, I am looking at this but it really takes quite some time for me understand this.  I have to say that the added logic for handling unaligned accesses makes the already complex assembly code even harder to read. And I am also worried about code maintainability as previously mentioned by @theRealAph about more checkings for AvoidUnalignedAccesses.

After a cursory look at this change, it seems that we could make use of the newly-added `load_X_misaligned` assemblers for the possibly unaligned memory accesses instead of adding the extra logic under AvoidUnalignedAccesses.  This might not offer us the optimal performance but it does reduce code complexity.

PS: At the same time, I think it's time to distinguish support for unaligned access for different platforms when measuring performance. We might want to manually set AvoidUnalignedAccesses to false for platforms like T-HEAD running older kernels. Thanks to: https://bugs.openjdk.org/browse/JDK-8309258, we have the proper detection for new kernels. We would add more vender-specific settings in the future.