@RealFYang 2023-06-26
> The compare_short ( in c2_macroAssember) was doing too much conditional branches in one place,it was possible to slightly reduce it. Thanks for looking at it.

I haven't checked other changes, but for changes made in file src/hotspot/cpu/riscv/c2_MacroAssembler_riscv.cpp:
https://github.com/openjdk/jdk/blob/master/src/hotspot/cpu/riscv/c2_MacroAssembler_riscv.cpp#LL958-L971

It looks that simple calls to load_X_misaligned assemblers with proper granularity for these unaligned accesses will suffice, which would be much simpler.